Refactor workload
