+++ workloads/imagenet/imagenet_jax/workload.py
@@ -23,14 +23,14 @@ from flax import jax_utils
 
 import spec
 import random_utils as prng
+from workloads.imagenet.workload import ImagenetWorkload
 from . import input_pipeline
 from . import models
 
 
-
-class ImagenetWorkload(spec.Workload):
+class ImagenetWorkload(ImagenetWorkload):
   def __init__(self):
-    self._eval_ds = None
+    super().__init__()
     self._param_shapes = None
     self.epoch_metrics = []
     # self.model_name = 'ResNet50'
@@ -40,28 +40,6 @@ class ImagenetWorkload(spec.Workload):
     self.model_name = '_ResNet1'
     self.dataset = 'imagenette'
     self.num_classes = 10
-
-  def has_reached_goal(self, eval_result: float) -> bool:
-    return eval_result['accuracy'] > self.target_value
-
-  @property
-  def target_value(self):
-    return 0.76
-
-  @property
-  def loss_type(self):
-    return spec.LossType.SOFTMAX_CROSS_ENTROPY
-
-  @property
-  def train_mean(self):
-    return [0.485 * 255, 0.456 * 255, 0.406 * 255]
-
-  @property
-  def train_stddev(self):
-    return [0.229 * 255, 0.224 * 255, 0.225 * 255]
-
-  def model_params_types(self):
-    pass
 
   @property
   def num_train_examples(self):
@@ -76,25 +54,6 @@ class ImagenetWorkload(spec.Workload):
       return 100000
     if 'imagenette' == self.dataset:
       return 3925
-
-  @property
-  def max_allowed_runtime_sec(self):
-    if 'imagenet2012' in self.dataset:
-      return 111600 # 31 hours
-    if 'imagenette' == self.dataset:
-      return 3600 # 60 minutes
-
-  @property
-  def eval_period_time_sec(self):
-    if 'imagenet2012' in self.dataset:
-      return 6000 # 100 mins
-    if 'imagenette' == self.dataset:
-      return 30 # 30 seconds
-
-  # Return whether or not a key in spec.ParameterContainer is the output layer
-  # parameters.
-  def is_output_params(self, param_key: spec.ParameterKey) -> bool:
-    pass
 
   def _build_dataset(self,
       data_rng: spec.RandomState,
@@ -113,14 +72,6 @@ class ImagenetWorkload(spec.Workload):
       train=True,
       cache=False)
     return ds
-
-  def build_input_queue(
-      self,
-      data_rng: spec.RandomState,
-      split: str,
-      data_dir: str,
-      batch_size: int):
-    return iter(self._build_dataset(data_rng, split, data_dir, batch_size))
 
   def sync_batch_stats(self, model_state):
     """Sync the batch statistics across replicas."""
@@ -262,5 +213,4 @@ class ImagenetWorkload(spec.Workload):
     eval_metrics = jax.tree_multimap(lambda *x: np.stack(x), *eval_metrics)
     summary = jax.tree_map(lambda x: x.mean(), eval_metrics)
     return summary
-
